Added transfer logic

+++ TeamCode/src/main/java/org/firstinspires/ftc/utils/DelayedActions.kt
@@ -4,13 +4,13 @@ package org.firstinspires.ftc.utils
  * A class for running actions after a delay.
  */
 class DelayedActions {
-    private val actions: MutableList<Pair<Long,() -> Unit>> = mutableListOf()
+    private val actions: MutableList<Pair<Double,() -> Unit>> = mutableListOf()
     /**
      * Adds an action to be run after a delay.
      * @param delay the delay in seconds.
      * @param action the action to run.
      */
-    fun addDelay(delay: Long, action: () -> Unit) {
+    fun addDelay(delay: Double, action: () -> Unit) {
         actions.add(Pair(delay + System.currentTimeMillis(), action))
     }
     /**
